+++ backend/canisters/multi_user/impl/src/queries/saved_crypto_accounts.rs
@@ -1,7 +1,13 @@
+use crate::guards::caller_is_hosted_user;
+use crate::{RuntimeState, read_state};
 use canister_api_macros::query;
-use user_canister::saved_crypto_accounts::*;
+use user_canister::saved_crypto_accounts::{Response::*, *};
 
-#[query(msgpack = true)]
+#[query(guard = "caller_is_hosted_user", msgpack = true)]
 fn saved_crypto_accounts(_args: Args) -> Response {
-    unimplemented!()
+    read_state(saved_crypto_accounts_impl)
+}
+
+fn saved_crypto_accounts_impl(state: &RuntimeState) -> Response {
+    Success(state.with_caller_user(|_, user| user.saved_crypto_accounts.all().to_vec()))
 }